+++ suit/tests/utils.py
@@ -2,11 +2,10 @@ from django import get_version
 from suit import utils
 from django.test import TestCase
 
+from suit.templatetags.suit_tags import str_to_version
+
 
 class UtilsTestCase(TestCase):
-    def test_django_major_version(self):
-        self.assertEqual(utils.django_major_version(), float(get_version()[:3]))
-
     def get_args(self):
         return [1.4, 'x', 1.5, 'y', 's', 'z']
 
@@ -19,3 +18,5 @@ class UtilsTestCase(TestCase):
         args = [utils.django_major_version(), 'a']
         self.assertEqual(utils.value_by_version(args), 'a')
 
+    def test_str_to_version(self):
+        return self.assertEqual(str_to_version('1.10.2'), (1, 10, 2))