+++ CHANGELOG.rst
@@ -4,6 +4,18 @@ Changelog
 Only important changes are mentioned below. See `commit log <https://github.com/darklow/django-suit/commits/develop>`_, `closed issues <https://github.com/darklow/django-suit/issues?direction=desc&sort=updated&state=closed>`_ and `closed pull
 requests <https://github.com/darklow/django-suit/pulls?q=sort%3Aupdated-desc+is%3Apr+is%3Aclosed>`_ for full changes.
 
+v0.2.21 (2016-08-18)
+--------------------
+
+* [Fix] Django 1.10 compatibility and Travis tests fixes
+
+
+v0.2.20 (2016-08-09)
+--------------------
+
+* [Fix] Django 1.10 compatibility fixes `#526 <https://github.com/darklow/django-suit/pull/526>`_
+
+
 v0.2.19 (2016-06-23)
 --------------------